backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.20
celery[redis]==5.4.0
redis==5.2.1
python-dateutil==2.9.0.post0
pypdf==5.3.0
python-docx==1.1.2
langdetect==1.0.9
reportlab==4.3.1
ffmpeg-python==0.2.0
PyYAML==6.0.2
httpx==0.28.1
pytest==8.3.4
pytest-cov==6.0.0